extract reader and writer
