+++ gql_code_gen/lib/builder.dart
@@ -1,16 +1,12 @@
-library ast_builder;
-
 import "dart:async";
 
 import "package:build/build.dart";
-import "package:code_builder/code_builder.dart";
-import "package:dart_style/dart_style.dart";
-import "package:gql/ast.dart";
-import "package:gql/language.dart";
+
 import "package:gql_code_gen/src/builders/ast.dart";
 import "package:gql_code_gen/src/builders/schema.dart";
-
-const schemaExtension = ".gql.dart";
+import "package:gql_code_gen/src/config.dart";
+import "package:gql_code_gen/src/reader/reader.dart";
+import "package:gql_code_gen/src/writer/writer.dart";
 
 /// Builder factory for Schema Builder
 Builder builder(
@@ -21,24 +17,17 @@ Builder builder(
 class _SchemaBuilder implements Builder {
   final AssetId schemaAsset;
 
-  final DartFormatter _dartfmt = DartFormatter();
-
   _SchemaBuilder(String schemaDescription)
       : schemaAsset = AssetId.parse(schemaDescription);
 
   @override
   Map<String, List<String>> get buildExtensions => {
-        ".graphql": [schemaExtension],
+        sourceExtension: [genExtension],
       };
-
-  Future<DocumentNode> readSchema(BuildStep buildStep) async => parseString(
-        await buildStep.readAsString(schemaAsset),
-        url: schemaAsset.uri.path,
-      );
 
   @override
   FutureOr<void> build(BuildStep buildStep) async {
-    final schema = await readSchema(buildStep);
+    final schema = await readDocument(buildStep, schemaAsset);
 
     if (buildStep.inputId == schemaAsset) {
       print("--------> [schema] = ${buildStep.inputId}");
@@ -49,29 +38,24 @@ class _SchemaBuilder implements Builder {
 
       if (library == null) return Future.value();
 
-      final genSrc = _dartfmt.format("${library.accept(
-        DartEmitter.scoped(),
-      )}");
-
-      return buildStep.writeAsString(
-        buildStep.inputId.changeExtension(schemaExtension),
-        genSrc,
+      return writeDocument(
+        library,
+        buildStep,
+        genExtension,
       );
     } else {
       print("--------> [not-schema] = ${buildStep.inputId}");
-      final src = await buildStep.readAsString(buildStep.inputId);
+      final doc = await readDocument(buildStep);
 
-      final doc = parseString(src, url: buildStep.inputId.path);
+      final library = buildAstLibrary(
+        doc,
+        schema,
+      );
 
-      final library = buildAstLibrary(doc, schema);
-
-      final genSrc = _dartfmt.format("${library.accept(
-        DartEmitter.scoped(),
-      )}");
-
-      return buildStep.writeAsString(
-        buildStep.inputId.changeExtension(schemaExtension),
-        genSrc,
+      return writeDocument(
+        library,
+        buildStep,
+        genExtension,
       );
     }
   }
