+++ gql_code_gen/lib/src/config.dart
@@ -1,0 +1,3 @@
+const sourceExtension = ".graphql";
+
+const genExtension = ".gql.dart";
