+++ gql_code_gen/lib/src/schema.dart
@@ -1,10 +1,12 @@
+import "package:meta/meta.dart";
 import "package:code_builder/code_builder.dart";
+
 import "package:gql/ast.dart";
+
 import "package:gql_code_gen/src/common.dart";
 import "package:gql_code_gen/src/schema/input.dart";
 import "package:gql_code_gen/src/schema/scalar.dart";
 import "package:gql_code_gen/src/schema/enum.dart";
-import "package:meta/meta.dart";
 
 /// Build input types, enums and scalars from schema
 Spec buildSchema(
@@ -79,5 +81,5 @@ class _SchemaBuilderVisitor extends SimpleVisitor<Spec> {
   Spec visitEnumTypeDefinitionNode(
     EnumTypeDefinitionNode node,
   ) =>
-    buildEnumClass(node);
+      buildEnumClass(node);
 }
